Added LC05111C13MTTTG as an alternate to AP9211SA-AG-HAC-7
I am still a little skeptical on the AP9211SA-AG-HAC-7’s discharge
current limit. LC05111C13MTTTG explicitly defines it as 3A, which is a
lot more reassuring than what the AP9211SA-AG-HAC-7 has to say about it.

For AP9211SA-AG-HAC-7, maximum discharge current was found by
V_overdischargelimit / R_SS(on). Gives an answer that works out with
power limit, but it’s still weird that the maximum discharge is not
given explicitly on data sheet (nor is the seemingly basic calculation
mentioned)